rebase

+++ fast.h
@@ -1,1 +1,1 @@
-long long int heap_ackermann_peter(long long int a, long long int b);
+int heap_ackermann_peter(int a, int b);
